gather_insights: if weblate created a PR, name all authors inside
Weblate creates a single PR that can include contributions from multiple
users.  The individual commits inside the PR have the actual contributor
as the "author".  Ensure that these authors are credited individually.

Sometimes no "author" is listed; this appears to be because the git
commit's author information does not match up to a github account.
In one case, I see the author information is listed as
"Anonymous <noreply@weblate.org>", which has no github user associated
and ends up uncredited.